Record selection of germline and C-region dbs in memory instead of file
Previous to this change igblastr was recording the selection of the germline
and C-region dbs in files located in the persistent user cache. This was
obviously a very bad idea because it meant that:
  - If the user was using igblastr in more than one R session, then all
    sessions were forced to use the same dbs.
  - Changing the selection in one session would change it for all other
    sessions, which was hugely problematic!

So now the db selection is recorded in memory rather than in a file. This
means that it is now remembered for the duration of the session only, and
does not persist across sessions. It also means that the user now needs to
specify the selection at the beginning of each session, which is actually
a good thing!